+++ src/components/footer-statistic.js
@@ -1,0 +1,7 @@
+const createFooterStatisticTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
+
+export {createFooterStatisticTemplate};
