+++ src/components/most-commented.js
@@ -1,0 +1,10 @@
+const createMostCommentedTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+        <h2 class="films-list__title">Most commented</h2>
+        <div class="films-list__container"></div>
+     </section>`
+  );
+};
+
+export {createMostCommentedTemplate};
