+++ src/components/show-more-button.js
@@ -1,0 +1,7 @@
+const createShowMoreButton = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createShowMoreButton};
