+++ src/components/top-rated.js
@@ -1,0 +1,10 @@
+const createTopRatedTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+        <h2 class="films-list__title">Top rated</h2>
+        <div class="films-list__container"></div>
+     </section>`
+  );
+};
+
+export {createTopRatedTemplate};
